pyproject.toml
[tool.poetry]
name = "siloka-ml-gateway"
version = "0.1.0"
description = ""
authors = ["Dennis A. Walangadi <dennisaw2000@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
fastapi = {extras = ["full"], version = "^0.78.0"}
numpy = "^1.22.4"
tensorflow = "^2.9.1"
python-dotenv = "^0.20.0"
uvicorn = "^0.17.6"

[tool.poetry.dev-dependencies]
pylint = "^2.14.0"
black = "^22.3.0"
setuptools = "^62.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
